requirement.txt
arabic-reshaper==2.1.3
asgiref==3.4.0
colorama==0.4.4
Django==2.2
django-crispy-forms==1.12.0
django-qr-code==2.2.0
future==0.18.2
html5lib==1.1
Pillow==8.2.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
qrcode==7.0
reportlab==3.5.68
segno==1.3.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
webencodings==0.5.1
xhtml2pdf==0.2.5